ref: Remove circular dependency in @sentry/node
